support test mode

+++ google/gstorage/gstorage.go
@@ -138,6 +138,30 @@ type Implementation struct {
 	projectID string
 }
 
+// testMode is true should return success, false return error, otherwise behave normal
+//
+var testMode *bool
+
+// TestModeAlwaySuccess will let every function success
+//
+func TestModeAlwaySuccess() {
+	t := true
+	testMode = &t
+}
+
+// TestModeAlwayFail will let every function fail
+//
+func TestModeAlwayFail() {
+	f := false
+	testMode = &f
+}
+
+// TestModeBackNormal stop test mode and back to normal
+//
+func TestModeBackNormal() {
+	testMode = nil
+}
+
 // New create Cloudstorage
 //
 //	storage, err := New(ctx,cred)
@@ -165,6 +189,12 @@ func New(ctx context.Context, cred *google.Credentials) (Gstorage, error) {
 //	err = storage.CreateBucket(ctx, "my.piyuo.com")
 //
 func (impl *Implementation) CreateBucket(ctx context.Context, bucketName string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
 
 	bucket := impl.client.Bucket(bucketName)
 	if err := bucket.Create(ctx, impl.projectID, &storage.BucketAttrs{
@@ -183,9 +213,14 @@ func (impl *Implementation) CreateBucket(ctx context.Context, bucketName string)
 //	err = storage.DeleteBucket(ctx, "my-bucket")
 //
 func (impl *Implementation) DeleteBucket(ctx context.Context, bucketName string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
 
 	bucket := impl.client.Bucket(bucketName)
-
 	if err := impl.CleanBucket(ctx, bucketName); err != nil {
 		return errors.Wrap(err, "clean bucket before remove "+bucketName)
 	}
@@ -203,6 +238,13 @@ func (impl *Implementation) DeleteBucket(ctx context.Context, bucketName string)
 //	err = storage.PublicBucket(ctx, "my-bucket")
 //
 func (impl *Implementation) PublicBucket(ctx context.Context, bucketName string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
+
 	bucket := impl.client.Bucket(bucketName)
 	policy, err := bucket.IAM().V3().Policy(ctx)
 	if err != nil {
@@ -225,6 +267,13 @@ func (impl *Implementation) PublicBucket(ctx context.Context, bucketName string)
 //	err = storage.SetPageAndCORS(ctx, "my-bucket","some-origin.com")
 //
 func (impl *Implementation) SetPageAndCORS(ctx context.Context, bucketName, originDomain string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
+
 	bucket := impl.client.Bucket(bucketName)
 	bucketAttrsToUpdate := storage.BucketAttrsToUpdate{
 		Website: &storage.BucketWebsite{
@@ -251,6 +300,13 @@ func (impl *Implementation) SetPageAndCORS(ctx context.Context, bucketName, orig
 //	exist, err := storage.IsBucketExists(ctx, bucketName)
 //
 func (impl *Implementation) IsBucketExists(ctx context.Context, bucketName string) (bool, error) {
+	if testMode != nil {
+		if *testMode {
+			return true, nil
+		}
+		return false, errors.New("fail")
+	}
+
 	bucketIterator := impl.client.Buckets(ctx, impl.projectID)
 	for {
 		bucketAttrs, err := bucketIterator.Next()
@@ -274,6 +330,13 @@ func (impl *Implementation) IsBucketExists(ctx context.Context, bucketName strin
 //	found,err = storage.IsFileExists(ctx, bucketName, "fileName")
 //
 func (impl *Implementation) IsFileExists(ctx context.Context, bucketName, path string) (bool, error) {
+	if testMode != nil {
+		if *testMode {
+			return true, nil
+		}
+		return false, errors.New("fail")
+	}
+
 	bucket := impl.client.Bucket(bucketName)
 	query := &storage.Query{}
 	it := bucket.Objects(ctx, query)
@@ -297,6 +360,13 @@ func (impl *Implementation) IsFileExists(ctx context.Context, bucketName, path s
 //	err = storage.WriteText(ctx, bucketName, "a/b.txt")
 //
 func (impl *Implementation) WriteText(ctx context.Context, bucketName, path, txt string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
+
 	bucket := impl.client.Bucket(bucketName)
 
 	wc := bucket.Object(path).NewWriter(ctx)
@@ -316,6 +386,13 @@ func (impl *Implementation) WriteText(ctx context.Context, bucketName, path, txt
 //	txt, err := storage.ReadText(ctx, bucketName, "a/b.txt")
 //
 func (impl *Implementation) ReadText(ctx context.Context, bucketName, path string) (string, error) {
+	if testMode != nil {
+		if *testMode {
+			return "", nil
+		}
+		return "", errors.New("fail")
+	}
+
 	bucket := impl.client.Bucket(bucketName)
 
 	rc, err := bucket.Object(path).NewReader(ctx)
@@ -336,6 +413,13 @@ func (impl *Implementation) ReadText(ctx context.Context, bucketName, path strin
 //	err = storage.WriteJSON(ctx, bucketName, "a/b.json")
 //
 func (impl *Implementation) WriteJSON(ctx context.Context, bucketName, path string, content map[string]interface{}) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
+
 	bytes, err := json.Marshal(content)
 	if err != nil {
 		return errors.Wrap(err, "marshal json")
@@ -351,6 +435,13 @@ func (impl *Implementation) WriteJSON(ctx context.Context, bucketName, path stri
 //	txt, err := storage.ReadJSON(ctx, bucketName, "a/b.json")
 //
 func (impl *Implementation) ReadJSON(ctx context.Context, bucketName, path string) (map[string]interface{}, error) {
+	if testMode != nil {
+		if *testMode {
+			return map[string]interface{}{}, nil
+		}
+		return map[string]interface{}{}, errors.New("fail")
+	}
+
 	var control map[string]interface{}
 	text, err := impl.ReadText(ctx, bucketName, path)
 	if err == nil {
@@ -367,6 +458,13 @@ func (impl *Implementation) ReadJSON(ctx context.Context, bucketName, path strin
 //	files, err := storage.ListFiles(ctx, bucketName, "", "")
 //
 func (impl *Implementation) ListFiles(ctx context.Context, bucketName, prefix, delim string) ([]string, error) {
+	if testMode != nil {
+		if *testMode {
+			return []string{}, nil
+		}
+		return []string{}, errors.New("fail")
+	}
+
 	files := []string{}
 	bucket := impl.client.Bucket(bucketName)
 	query := &storage.Query{Prefix: prefix, Delimiter: delim}
@@ -392,6 +490,13 @@ func (impl *Implementation) ListFiles(ctx context.Context, bucketName, prefix, d
 //	err = storage.DeleteFiles(ctx, bucketName, "assets")
 //
 func (impl *Implementation) DeleteFiles(ctx context.Context, bucketName, prefix string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
+
 	query := &storage.Query{Prefix: prefix}
 	query.SetAttrSelection([]string{"Name"})
 	it := impl.client.Bucket(bucketName).Objects(ctx, query)
@@ -418,6 +523,12 @@ func (impl *Implementation) DeleteFiles(ctx context.Context, bucketName, prefix 
 //	err = storage.DeleteFile(ctx, bucketName, path)
 //
 func (impl *Implementation) DeleteFile(ctx context.Context, bucketName, path string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
 
 	o := impl.client.Bucket(bucketName).Object(path)
 	if err := o.Delete(ctx); err != nil {
@@ -434,6 +545,13 @@ func (impl *Implementation) DeleteFile(ctx context.Context, bucketName, path str
 //	err = storage.Delete(ctx, bucketName, path)
 //
 func (impl *Implementation) CleanBucket(ctx context.Context, bucketName string) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
+
 	bucket := impl.client.Bucket(bucketName)
 	for {
 		result, err := impl.removeObjects(ctx, bucket)
@@ -449,6 +567,13 @@ func (impl *Implementation) CleanBucket(ctx context.Context, bucketName string) 
 // removeObjects remove objects max 1000, return true if object all deleted
 //
 func (impl *Implementation) removeObjects(ctx context.Context, bucket *storage.BucketHandle) (bool, error) {
+	if testMode != nil {
+		if *testMode {
+			return true, nil
+		}
+		return false, errors.New("fail")
+	}
+
 	query := &storage.Query{}
 	query.SetAttrSelection([]string{"Name"})
 
@@ -487,6 +612,12 @@ type ShouldDeleteFile func(filename string) (bool, error)
 //	err = storage.Sync(ctx, bucketName, dir)
 //
 func (impl *Implementation) Sync(ctx context.Context, bucketName string, shouldDeleteFile ShouldDeleteFile) error {
+	if testMode != nil {
+		if *testMode {
+			return nil
+		}
+		return errors.New("fail")
+	}
 
 	bucket := impl.client.Bucket(bucketName)
 	query := &storage.Query{}
